move `/` to `/sessions`
