+++ src/routes/sessions/[id]/+page.svelte
