+++ src/routes/sessions/[id]/+page.ts
