+++ src/routes/sessions/[id]/CopyButton.svelte
